backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
bcrypt==4.1.2
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0
requests==2.31.0
gunicorn==21.2.0
Werkzeug==3.0.1